Fix : Nix installation command by correctly using the -L flag in curl command
correction of  the usage of the -L flag in the Nix installation command, ensuring that curl handles redirects properly. The -L flag is now correctly placed with a space separating it from the URL to avoid argument misinterpretation. This change resolves an error where curl was not following redirects as intended, improving the reliability of the installation command for all users